+++ packages/plugin-vue/src/index.ts
@@ -14,7 +14,9 @@ import {
   SFCBlock,
   SFCScriptCompileOptions,
   SFCStyleCompileOptions,
-  SFCTemplateCompileOptions
+  SFCTemplateCompileOptions,
+  shouldTransformRef,
+  transformRef
 } from '@vue/compiler-sfc'
 import { parseVueRequest } from './utils/query'
 import { getDescriptor } from './utils/descriptorCache'
@@ -55,6 +57,22 @@ export interface Options {
   customElement?: boolean | string | RegExp | (string | RegExp)[]
 
   /**
+   * Enable Vue ref transform (experimental).
+   * https://github.com/vuejs/vue-next/tree/master/packages/ref-transform
+   *
+   * **requires Vue \>= 3.2.5**
+   *
+   * - `true`: transform will be enabled for all vue,js(x),ts(x) files except
+   *           those inside node_modules
+   * - `string | RegExp`: apply to vue + only matched files (will include
+   *                      node_modules, so specify directories in necessary)
+   * - `false`: disable in all cases
+   *
+   * @default false
+   */
+  refTransform?: boolean | string | RegExp | (string | RegExp)[]
+
+  /**
    * @deprecated the plugin now auto-detects whether it's being invoked for ssr.
    */
   ssr?: boolean
@@ -66,21 +84,39 @@ export interface ResolvedOptions extends Options {
 }
 
 export default function vuePlugin(rawOptions: Options = {}): Plugin {
+  const {
+    include = /\.vue$/,
+    exclude,
+    customElement = /\.ce\.vue$/,
+    refTransform = false
+  } = rawOptions
+
+  const filter = createFilter(include, exclude)
+
+  const customElementFilter =
+    typeof customElement === 'boolean'
+      ? () => customElement
+      : createFilter(customElement)
+
+  const refTransformFilter =
+    refTransform === false
+      ? () => false
+      : refTransform === true
+      ? createFilter(/\.(j|t)sx?$/, /node_modules/)
+      : createFilter(refTransform)
+
+  // compat for older verisons
+  const canUseRefTransform = typeof shouldTransformRef === 'function'
+
   let options: ResolvedOptions = {
     isProduction: process.env.NODE_ENV === 'production',
     ...rawOptions,
+    include,
+    exclude,
+    customElement,
+    refTransform,
     root: process.cwd()
   }
-
-  const filter = createFilter(
-    rawOptions.include || /\.vue$/,
-    rawOptions.exclude
-  )
-
-  const customElementFilter =
-    typeof rawOptions.customElement === 'boolean'
-      ? () => rawOptions.customElement as boolean
-      : createFilter(rawOptions.customElement || /\.ce\.vue$/)
 
   return {
     name: 'vite:vue',
@@ -131,7 +167,11 @@ export default function vuePlugin(rawOptions: Options = {}): Plugin {
         if (query.src) {
           return fs.readFileSync(filename, 'utf-8')
         }
-        const descriptor = getDescriptor(filename)!
+        const descriptor = getDescriptor(
+          filename,
+          options.root,
+          options.isProduction
+        )!
         let block: SFCBlock | null | undefined
         if (query.type === 'script') {
           // handle <scrip> + <script setup> merge via compileScript()
@@ -154,7 +194,20 @@ export default function vuePlugin(rawOptions: Options = {}): Plugin {
 
     transform(code, id, ssr = !!options.ssr) {
       const { filename, query } = parseVueRequest(id)
-      if ((!query.vue && !filter(filename)) || query.raw) {
+      if (query.raw) {
+        return
+      }
+      if (!filter(filename) && !query.vue) {
+        if (!query.vue && refTransformFilter(filename)) {
+          if (!canUseRefTransform) {
+            this.warn('refTransform requires @vue/compiler-sfc@^3.2.5.')
+          } else if (shouldTransformRef(code)) {
+            return transformRef(code, {
+              filename,
+              sourceMap: true
+            })
+          }
+        }
         return
       }
 
@@ -170,7 +223,11 @@ export default function vuePlugin(rawOptions: Options = {}): Plugin {
         )
       } else {
         // sub block request
-        const descriptor = getDescriptor(filename)!
+        const descriptor = getDescriptor(
+          filename,
+          options.root,
+          options.isProduction
+        )!
         if (query.type === 'template') {
           return transformTemplateAsModule(code, descriptor, options, this, ssr)
         } else if (query.type === 'style') {