+++ packages/vite/client.d.ts
@@ -161,6 +161,10 @@ declare module '*.webmanifest' {
   const src: string
   export default src
 }
+declare module '*.pdf' {
+  const src: string
+  export default src
+}
 
 // web worker
 declare module '*?worker' {