+++ packages/vite/src/node/optimizer/scan.ts
@@ -15,7 +15,9 @@ import {
   isObject,
   cleanUrl,
   externalRE,
-  dataUrlRE
+  dataUrlRE,
+  multilineCommentsRE,
+  singlelineCommentsRE
 } from '../utils'
 import {
   createPluginContainer,
@@ -39,9 +41,6 @@ const htmlTypesRE = /\.(html|vue|svelte)$/
 // simply be ignored.
 export const importsRE =
   /(?<!\/\/.*)(?<=^|;|\*\/)\s*import(?!\s+type)(?:[\w*{}\n\r\t, ]+from\s*)?\s*("[^"]+"|'[^']+')\s*(?=$|;|\/\/|\/\*)/gm
-
-export const multilineCommentsRE = /\/\*(.|[\r\n])*?\*\//gm
-export const singlelineCommentsRE = /\/\/.*/g
 
 export async function scanImports(config: ResolvedConfig): Promise<{
   deps: Record<string, string>
@@ -163,7 +162,11 @@ function esbuildScanPlugin(
   }
 
   const include = config.optimizeDeps?.include
-  const exclude = config.optimizeDeps?.exclude
+  const exclude = [
+    ...(config.optimizeDeps?.exclude || []),
+    '@vite/client',
+    '@vite/env'
+  ]
 
   const externalUnlessEntry = ({ path }: { path: string }) => ({
     path,