+++ packages/create-vite/CHANGELOG.md
@@ -1,3 +1,36 @@
+## [2.6.2](https://github.com/vitejs/vite/compare/create-vite@2.6.1...create-vite@2.6.2) (2021-09-07)
+
+
+
+## [2.6.1](https://github.com/vitejs/vite/compare/create-vite@2.5.4...create-vite@2.6.1) (2021-08-31)
+
+
+### Bug Fixes
+
+* **create-vite:** project name with only numbers as an argument ([#4606](https://github.com/vitejs/vite/issues/4606)) ([085a621](https://github.com/vitejs/vite/commit/085a6219ceb89052ac0a269e7ccda081e534a800))
+
+
+### Features
+
+* **create-vite:** update vue templates ([87a9a70](https://github.com/vitejs/vite/commit/87a9a70f75991a8afb06a0b84f4eab43ca76fa89))
+* explicit set `useDefineForClassFields` in ts templates ([#4280](https://github.com/vitejs/vite/issues/4280)) ([fe74173](https://github.com/vitejs/vite/commit/fe74173f19b6e341114723a99043440abce27bd6))
+
+
+
+# [2.6.0](https://github.com/vitejs/vite/compare/create-vite@2.5.4...create-vite@2.6.0) (2021-08-25)
+
+
+### Bug Fixes
+
+* **create-vite:** project name with only numbers as an argument ([#4606](https://github.com/vitejs/vite/issues/4606)) ([085a621](https://github.com/vitejs/vite/commit/085a6219ceb89052ac0a269e7ccda081e534a800))
+
+
+### Features
+
+* **create-vite:** update vue templates ([fe9ebe7](https://github.com/vitejs/vite/commit/fe9ebe7bb227979e9420ba295b892a716daaaf69))
+
+
+
 ## [2.5.4](https://github.com/vitejs/vite/compare/create-vite@2.5.3...create-vite@2.5.4) (2021-08-03)
 
 
@@ -43,7 +76,7 @@ $ yarn create vite
 With PNPM:
 
 ```bash
-$ pnpx create-vite
+$ pnpm dlx create-vite
 ```